Rename Solution, Project and Merg code till 20/02/2025

+++ AppraisalBackend/DBL/Models/TransactionLog.cs
@@ -25,9 +25,9 @@
 
         public DateTime? EndDate { get; set; }
 
-        public short? UsedProperties { get; set; }
+        public long? UsedProperties { get; set; }
 
-        public short TotalProperties { get; set; }
+        public long TotalProperties { get; set; }
 
         public string? PaypalSubscriptionStatus { get; set; }
     }
